requirements.txt
streamlit>=1.28.0
groq>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
gtts>=2.3.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
